+++ src/components/RequestManager/ManageRequestCampaignsPage/Feature/EditStatusForm.js
@@ -1,0 +1,104 @@
+import { Button } from "../../../ui/button";
+import {
+  Dialog,
+  DialogClose,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "../../../ui/dialog";
+import { useFormik } from "formik";
+import { useToast } from "../../../ui/use-toast";
+import { Label } from "../../../ui/label";
+import { Input } from "../../../ui/input";
+import { CopyButton } from "./CopyButton";
+import { Avatar, AvatarFallback, AvatarImage } from "../../../ui/avatar";
+import { Switch } from "../../../ui/switch";
+import { Badge } from "../../../ui/badge";
+import React from "react";
+const EditStatusForm = ({ isOpen, onOpenChange, campaign }) => {
+  const { toast } = useToast();
+  // Formik setup
+  const formik = useFormik({
+    initialValues: {
+      is_approved: campaign ? campaign.is_approved : false,
+
+    },
+    onSubmit: (values, { setSubmitting }) => {
+      toast({
+        title: "Thông tin duyệt:",
+        description: (
+          <pre className="mt-2 w-[340px] rounded-md bg-black p-4">
+            <code className="text-white">
+              {JSON.stringify(values, null, 2)}
+            </code>{" "}
+            {/* For testing*/}
+          </pre>
+        ),
+      });
+      setSubmitting(false);
+      onOpenChange(false); // Close the dialog after form submission
+    },
+  });
+  /* Giải thích: 
+  Vấn đề ở đây là formik là một đối tượng được tạo ra bởi hook useFormik, và nó thay đổi mỗi khi component re-render. Khi mình thêm formik vào mảng dependencies của useEffect, nó sẽ chạy mỗi khi formik thay đổi, tức là mỗi khi component re-render. Một cách để giải quyết vấn đề này là sử dụng useRef để lưu trữ giá trị formik.setValues và sau đó sử dụng giá trị đó trong useEffect.
+   */
+  const setValuesRef = React.useRef(formik.setValues);
+  // Update formik initialValues when campaign changes
+  React.useEffect(() => {
+    setValuesRef.current({
+      is_approved: campaign ? campaign.is_approved : false,
+
+    });
+  }, [campaign]);
+  // Handle switch change
+  const handleSwitchChange = (field) => (isChecked) => {
+    formik.setFieldValue(field, isChecked);
+  };
+
+  return (
+    <Dialog open={isOpen} onOpenChange={onOpenChange}>
+      <DialogContent className="mobile:max-w-md flex flex-col">
+        <DialogHeader>
+          <DialogTitle>Tổng duyệt trạng thái</DialogTitle>
+          <DialogDescription>
+            Lưu ý: Bạn chỉ có thể chỉnh sửa trạng thái xác thực của chiến dịch!
+          </DialogDescription>
+        </DialogHeader>
+     
+     
+        {campaign && (
+          <form onSubmit={formik.handleSubmit} className="space-y-3">
+            {/*  */}
+            <div className="flex items-center space-x-2">
+              <Switch
+                id="is_approved"
+                checked={formik.values.is_approved}
+                onCheckedChange={handleSwitchChange("is_approved")}
+              />
+              <Label htmlFor="is_approved">Chấp thuận</Label>
+            </div>
+            
+          </form>
+        )}
+        <DialogFooter>
+          <DialogClose asChild>
+            <Button type="button" variant="secondary">
+              Đóng
+            </Button>
+          </DialogClose>
+          <Button
+            type="button"
+            disabled={formik.isSubmitting}
+            onClick={formik.handleSubmit}
+          >
+            Xác nhận
+          </Button>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
+  );
+};
+
+export default EditStatusForm;
